+++ index.js
@@ -1,7 +1,14 @@
 
 const Layer = require('express/lib/router/layer')
+const flatten = require('array-flatten')
 
-module.exports = (path, middleware, options = {}) => {
+module.exports = (path, ...middleware) => {
+  // if first middleware is an object, it's options
+  const options = {}
+  if (middleware[0] && typeof middleware[0] === 'object' && !Array.isArray(middleware[0])) {
+    Object.assign(options, middleware.shift())
+  }
+
   const layerOpts = {
     sensitive: 'caseSensitive' in options ? options.caseSensitive : false,
     strict: 'strict' in options ? options.strict : false,
@@ -14,10 +21,7 @@ module.exports = (path, middleware, options = {}) => {
 
   const layer = new Layer(path, layerOpts, noOp)
 
-  if (!Array.isArray(middleware)) {
-    middleware = [ middleware ]
-  }
-  return middleware.map(_middleware => (req, res, next) => {
+  return flatten(middleware).map(_middleware => (req, res, next) => {
     if (layer.match(req.path)) return next()
     _middleware(req, res, next)
   })